+++ CHANGELOG.md
@@ -4,6 +4,10 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## Unreleased
 
+- Fix typo in `docker.up` and `docker.up_containers` parameters:
+  - `quite_pull` -> `quiet_pull`
+  - `quite_build` -> `quiet_build`
+
 ## 1.12.1
 
 - Remove `kwargs` from invocations in favor of explicit parameters. `kwargs` not